refactor: moved Vec impl Cow into cow.rs
